ThumbanilStore 구조 변경 + 예외 Advice 처리 + API 명세서 업데이트
